[57] Added resize to browser test. solves issue of not being able to click button.
