+++ .travis.yml
@@ -17,5 +17,3 @@ before_script:
   - 'sh -e /etc/init.d/xvfb start'
   - './vendor/laravel/dusk/bin/chromedriver-linux &'
   - 'php artisan serve &'
-  script:
-  - php artisan dusk
